pyproject.toml
[tool.poetry]
name = "accessibility-events"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "accessibility_events"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
peewee = "^3.16.3"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
whyslow = "^0.1.7"
selenium = "^4.14.0"

[tool.poetry.scripts]
dev = "accessibility_events.main:main"
test = "accessibility_events.test:main"
emails = "accessibility_events.emails:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
